Move preempted_workloads to preemption package
